Refactored Breadcrumb so it has extracted renderCrumbs + no longer needs boolean property
