+++ pages/docs/text-inputs-article.tsx
@@ -1,5 +1,5 @@
 import React, { useState, FC } from "react"
-import Breadcrumbs from "@designSystem/layouts/Breadcrumbs"
+import Breadcrumbs from "@designSystem/molecules/Breadcrumbs"
 import { useForm } from "react-hook-form"
 import WrapperInput from "@forms/WrapperInput"
 import LayoutContainerSide from "@designSystem/layouts/LayoutContainerSide"
