+++ tsconfig.json
@@ -36,7 +36,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "designSystem/layouts/Breadcrumbs.tsx"
+    "designSystem/molecules/Breadcrumbs.tsx"
   ],
   "exclude": ["node_modules"]
 }
